# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Projecte Integrador API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/google/redirect": {
            "get": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Redirige a Google para iniciar sesión",
                "description": "Redirige al usuario a la página de inicio de sesión de Google.",
                "operationId": "b866c3b42cb895103d6d88e22b7a5c72",
                "responses": {
                    "302": {
                        "description": "Redirección a Google"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/auth/google/callback": {
            "get": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Gestiona la respuesta de Google tras el inicio de sesión",
                "description": "Obtiene la información del usuario de Google tras el inicio de sesión y genera un token.",
                "operationId": "210885a42fb8000ebe58f4de134154c5",
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "description": "Token de autenticación",
                                            "type": "string"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Usuario no encontrado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Inicia sesión con credenciales",
                "description": "Inicia sesión con un email y contraseña.",
                "operationId": "8bb20529d05d50d3c283a0dcfe716bfe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "usuario@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "description": "Token de autenticación",
                                            "type": "string"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales inválidas"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cierra la sesión del usuario",
                "description": "Cierra la sesión del usuario actual, invalidando el token.",
                "operationId": "df9bdc6c7a6d9f20b983526b2ec0f698",
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User successfully signed out."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "User name",
                        "type": "string"
                    },
                    "email": {
                        "description": "User email",
                        "type": "string",
                        "format": "email"
                    },
                    "telefono": {
                        "description": "User phone number",
                        "type": "string"
                    },
                    "lenguas": {
                        "description": "Languages spoken by the user",
                        "type": "string"
                    },
                    "fecha_contratacion": {
                        "description": "User hire date",
                        "type": "string",
                        "format": "date"
                    },
                    "fecha_baja": {
                        "description": "User termination date",
                        "type": "string",
                        "format": "date"
                    },
                    "username": {
                        "description": "User username",
                        "type": "string"
                    },
                    "apellidos": {
                        "description": "User last names",
                        "type": "string"
                    },
                    "role": {
                        "description": "User role",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}